+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - LACLI006 - 21/04/2023 - Add the AESCommunicator class.
 - LACLI006 - 21/04/2023 - Add the RSACommunicator class.
 - LACLI006 - 21/04/2023 - Add the PlainCommunicator class.
+- LACLI006 - 21/04/2023 - Add the AESGenerator class.
 
 ### Changed